Moved Next and Prev buttons to ProjectsIndex (parent component)
